update to fossil-src-2.6

+++ src/name.c
@@ -4,7 +4,7 @@
 ** This program is free software; you can redistribute it and/or
 ** modify it under the terms of the Simplified BSD License (also
 ** known as the "2-Clause License" or "FreeBSD License".)
-
+**
 ** This program is distributed in the hope that it will be useful,
 ** but without any warranty; without even the implied warranty of
 ** merchantability or fitness for a particular purpose.
@@ -15,11 +15,7 @@
 **
 *******************************************************************************
 **
-** This file contains code used to convert user-supplied object names into
-** canonical UUIDs.
-**
-** A user-supplied object name is any unique prefix of a valid UUID but
-** not necessarily in canonical form.
+** This file contains code used to resolved user-supplied object names.
 */
 #include "config.h"
 #include "name.h"
@@ -82,8 +78,8 @@ int start_of_branch(int rid, int inBranch){
 /*
 ** Convert a symbolic name into a RID.  Acceptable forms:
 **
-**   *  SHA1 hash
-**   *  SHA1 hash prefix of at least 4 characters
+**   *  artifact hash (optionally enclosed in [...])
+**   *  4-character or larger prefix of a artifact
 **   *  Symbolic Name
 **   *  "tag:" + symbolic name
 **   *  Date or date-time
@@ -107,7 +103,7 @@ int start_of_branch(int rid, int inBranch){
 ** zType is "ci" in most use cases since we are usually searching for
 ** a check-in.
 **
-** Note that the input zTag for types "t" and "e" is the SHA1 hash of
+** Note that the input zTag for types "t" and "e" is the artifact hash of
 ** the ticket-change or event-change artifact, not the randomly generated
 ** hexadecimal identifier assigned to tickets and events.  Those identifiers
 ** live in a separate namespace.
@@ -118,6 +114,8 @@ int symbolic_name_to_rid(const char *zTag, const char *zType){
   int nTag;
   int i;
   int startOfBranch = 0;
+  const char *zXTag;     /* zTag with optional [...] removed */
+  int nXTag;             /* Size of zXTag */
 
   if( zType==0 || zType[0]==0 ){
     zType = "*";
@@ -156,7 +154,7 @@ int symbolic_name_to_rid(const char *zTag, const char *zType){
   if( memcmp(zTag, "date:", 5)==0 ){
     rid = db_int(0,
       "SELECT objid FROM event"
-      " WHERE mtime<=julianday(%Q,'utc') AND type GLOB '%q'"
+      " WHERE mtime<=julianday(%Q,fromLocal()) AND type GLOB '%q'"
       " ORDER BY mtime DESC LIMIT 1",
       &zTag[5], zType);
     return rid;
@@ -164,7 +162,7 @@ int symbolic_name_to_rid(const char *zTag, const char *zType){
   if( fossil_isdate(zTag) ){
     rid = db_int(0,
       "SELECT objid FROM event"
-      " WHERE mtime<=julianday(%Q,'utc') AND type GLOB '%q'"
+      " WHERE mtime<=julianday(%Q,fromLocal()) AND type GLOB '%q'"
       " ORDER BY mtime DESC LIMIT 1",
       zTag, zType);
     if( rid) return rid;
@@ -234,12 +232,24 @@ int symbolic_name_to_rid(const char *zTag, const char *zType){
     return rid;
   }
 
-  /* SHA1 hash or prefix */
-  if( nTag>=4 && nTag<=UUID_SIZE && validate16(zTag, nTag) ){
+  /* Remove optional [...] */
+  zXTag = zTag;
+  nXTag = nTag;
+  if( zXTag[0]=='[' ){
+    zXTag++;
+    nXTag--;
+  }
+  if( nXTag>0 && zXTag[nXTag-1]==']' ){
+    nXTag--;
+  }
+
+  /* artifact hash or prefix */
+  if( nXTag>=4 && nXTag<=HNAME_MAX && validate16(zXTag, nXTag) ){
     Stmt q;
-    char zUuid[UUID_SIZE+1];
-    memcpy(zUuid, zTag, nTag+1);
-    canonical16(zUuid, nTag);
+    char zUuid[HNAME_MAX+1];
+    memcpy(zUuid, zXTag, nXTag);
+    zUuid[nXTag] = 0;
+    canonical16(zUuid, nXTag);
     rid = 0;
     if( zType[0]=='*' ){
       db_prepare(&q, "SELECT rid FROM blob WHERE uuid GLOB '%q*'", zUuid);
@@ -335,7 +345,7 @@ int name_to_uuid(Blob *pName, int iErrPriority, const char *zType){
 ** name_to_uuid(). zType is also as described for that function. If
 ** zName does not resolve, 0 is returned. If it is ambiguous, a
 ** negative value is returned. On success the rid is returned and
-** pUuid (if it is not NULL) is set to the a newly-allocated string,
+** pUuid (if it is not NULL) is set to a newly-allocated string,
 ** the full UUID, which must eventually be free()d by the caller.
 */
 int name_to_uuid2(const char *zName, const char *zType, char **pUuid){
@@ -356,7 +366,7 @@ int name_collisions(const char *zName){
   int c = 0;         /* count of collisions for zName */
   int nLen;          /* length of zName */
   nLen = strlen(zName);
-  if( nLen>=4 && nLen<=UUID_SIZE && validate16(zName, nLen) ){
+  if( nLen>=4 && nLen<=HNAME_MAX && validate16(zName, nLen) ){
     c = db_int(0,
       "SELECT"
       " (SELECT count(*) FROM ticket"
@@ -373,7 +383,7 @@ int name_collisions(const char *zName){
 }
 
 /*
-** COMMAND:  test-name-to-id
+** COMMAND: test-name-to-id
 **
 ** Convert a name to a full artifact ID.
 */
@@ -398,7 +408,7 @@ void test_name_to_id(void){
 ** Convert a name to a rid.  If the name can be any of the various forms
 ** accepted:
 **
-**   * SHA1 hash or prefix thereof
+**   * artifact hash or prefix thereof
 **   * symbolic name
 **   * date
 **   * label:date
@@ -427,9 +437,9 @@ int name_to_rid(const char *zName){
 
 /*
 ** WEBPAGE: ambiguous
-** URL: /ambiguous?name=UUID&src=WEBPAGE
+** URL: /ambiguous?name=NAME&src=WEBPAGE
 **
-** The UUID given by the name parameter is ambiguous.  Display a page
+** The NAME given by the name parameter is ambiguous.  Display a page
 ** that shows all possible choices and let the user select between them.
 */
 void ambiguous_page(void){
@@ -473,7 +483,7 @@ void ambiguous_page(void){
     @ <ul></ul>
     @ Ticket
     hyperlink_to_uuid(zUuid);
-    @ - %s(zTitle).
+    @ - %h(zTitle).
     @ <ul><li>
     object_description(rid, 0, 0);
     @ </li></ul>
@@ -532,11 +542,11 @@ void whatis_rid(int rid, int verboseFlag){
 
   /* Basic information about the object. */
   db_prepare(&q,
-     "SELECT uuid, size, datetime(mtime%s), ipaddr"
+     "SELECT uuid, size, datetime(mtime,toLocal()), ipaddr"
      "  FROM blob, rcvfrom"
      " WHERE rid=%d"
      "   AND rcvfrom.rcvid=blob.rcvid",
-     timeline_utc(), rid);
+     rid);
   if( db_step(&q)==SQLITE_ROW ){
     if( verboseFlag ){
       fossil_print("artifact:   %s (%d)\n", db_column_text(&q,0), rid);
@@ -587,9 +597,9 @@ void whatis_rid(int rid, int verboseFlag){
 
   /* Check for entries on the timeline that reference this object */
   db_prepare(&q,
-     "SELECT type, datetime(mtime%s),"
+     "SELECT type, datetime(mtime,toLocal()),"
      "       coalesce(euser,user), coalesce(ecomment,comment)"
-     "  FROM event WHERE objid=%d", timeline_utc(), rid);
+     "  FROM event WHERE objid=%d", rid);
   if( db_step(&q)==SQLITE_ROW ){
     const char *zType;
     switch( db_column_text(&q,0)[0] ){
@@ -609,7 +619,7 @@ void whatis_rid(int rid, int verboseFlag){
 
   /* Check to see if this object is used as a file in a check-in */
   db_prepare(&q,
-    "SELECT filename.name, blob.uuid, datetime(event.mtime%s),"
+    "SELECT filename.name, blob.uuid, datetime(event.mtime,toLocal()),"
     "       coalesce(euser,user), coalesce(ecomment,comment)"
     "  FROM mlink, filename, blob, event"
     " WHERE mlink.fid=%d"
@@ -617,7 +627,7 @@ void whatis_rid(int rid, int verboseFlag){
     "   AND event.objid=mlink.mid"
     "   AND blob.rid=mlink.mid"
     " ORDER BY event.mtime DESC /*sort*/",
-    timeline_utc(), rid);
+    rid);
   while( db_step(&q)==SQLITE_ROW ){
     fossil_print("file:       %s\n", db_column_text(&q,0));
     fossil_print("            part of [%S] by %s on %s\n",
@@ -634,7 +644,7 @@ void whatis_rid(int rid, int verboseFlag){
     "SELECT attachment.filename,"
     "       attachment.comment,"
     "       attachment.user,"
-    "       datetime(attachment.mtime%s),"
+    "       datetime(attachment.mtime,toLocal()),"
     "       attachment.target,"
     "       CASE WHEN EXISTS(SELECT 1 FROM tag WHERE tagname=('tkt-'||target))"
     "            THEN 'ticket'"
@@ -644,7 +654,7 @@ void whatis_rid(int rid, int verboseFlag){
     "       (SELECT uuid FROM blob WHERE rid=attachid)"
     "  FROM attachment JOIN blob ON attachment.src=blob.uuid"
     " WHERE blob.rid=%d",
-    timeline_utc(), rid
+    rid
   );
   while( db_step(&q)==SQLITE_ROW ){
     fossil_print("attachment: %s\n", db_column_text(&q,0));
@@ -667,11 +677,18 @@ void whatis_rid(int rid, int verboseFlag){
 
 /*
 ** COMMAND: whatis*
+**
 ** Usage: %fossil whatis NAME
 **
-** Resolve the symbol NAME into its canonical 40-character SHA1-hash
+** Resolve the symbol NAME into its canonical artifact hash
 ** artifact name and provide a description of what role that artifact
 ** plays.
+**
+** Options:
+**
+**    --type TYPE          Only find artifacts of TYPE (one of: 'ci', 't',
+**                         'w', 'g', or 'e').
+**    -v|--verbose         Provide extra information (such as the RID)
 */
 void whatis_cmd(void){
   int rid;
@@ -686,7 +703,7 @@ void whatis_cmd(void){
   /* We should be done with options.. */
   verify_all_options();
 
-  if( g.argc<3 ) usage("whatis NAME ...");
+  if( g.argc<3 ) usage("NAME ...");
   for(i=2; i<g.argc; i++){
     zName = g.argv[i];
     if( i>2 ) fossil_print("%.79c\n",'-');
@@ -716,6 +733,7 @@ void whatis_cmd(void){
 
 /*
 ** COMMAND: test-whatis-all
+**
 ** Usage: %fossil test-whatis-all
 **
 ** Show "whatis" information about every artifact in the repository
@@ -735,9 +753,10 @@ void test_whatis_all_cmd(void){
 
 /*
 ** COMMAND: test-ambiguous
+**
 ** Usage: %fossil test-ambiguous [--minsize N]
 **
-** Show a list of ambiguous SHA1-hash abbreviations of N characters or
+** Show a list of ambiguous artifact hash abbreviations of N characters or
 ** more where N defaults to 4.  Change N to a different value using
 ** the "--minsize N" command-line option.
 */
@@ -787,7 +806,7 @@ void test_ambiguous_cmd(void){
 static const char zDescTab[] =
 @ CREATE TEMP TABLE IF NOT EXISTS description(
 @   rid INTEGER PRIMARY KEY,       -- RID of the object
-@   uuid TEXT,                     -- SHA1 hash of the object
+@   uuid TEXT,                     -- hash of the object
 @   ctime DATETIME,                -- Time of creation
 @   isPrivate BOOLEAN DEFAULT 0,   -- True for unpublished artifacts
 @   type TEXT,                     -- file, checkin, wiki, ticket, etc.
@@ -841,11 +860,12 @@ void describe_artifacts(const char *zWhere){
   /* Cluster artifacts */
   db_multi_exec(
     "INSERT OR IGNORE INTO description(rid,uuid,ctime,type,summary)\n"
-    "SELECT blob.rid, blob.uuid, tagxref.mtime, 'cluster', 'cluster'\n"
-    "  FROM tagxref, blob\n"
+    "SELECT blob.rid, blob.uuid, rcvfrom.mtime, 'cluster', 'cluster'\n"
+    "  FROM tagxref, blob, rcvfrom\n"
     " WHERE (tagxref.rid %s)\n"
     "   AND tagxref.tagid=(SELECT tagid FROM tag WHERE tagname='cluster')\n"
-    "   AND blob.rid=tagxref.rid;",
+    "   AND blob.rid=tagxref.rid"
+    "   AND rcvfrom.rcvid=blob.rcvid;",
     zWhere /*safe-for-%s*/
   );
 
@@ -926,12 +946,17 @@ void describe_artifacts(const char *zWhere){
 }
 
 /*
-** Print the content of the description table on stdout
+** Print the content of the description table on stdout.
+**
+** The description table is computed using the WHERE clause zWhere if
+** the zWhere parameter is not NULL.  If zWhere is NULL, then this
+** routine assumes that the description table already exists and is
+** populated and merely prints the contents.
 */
 int describe_artifacts_to_stdout(const char *zWhere, const char *zLabel){
   Stmt q;
   int cnt = 0;
-  describe_artifacts(zWhere);
+  if( zWhere!=0 ) describe_artifacts(zWhere);
   db_prepare(&q,
     "SELECT uuid, summary, isPrivate\n"
     "  FROM description\n"
@@ -948,7 +973,7 @@ int describe_artifacts_to_stdout(const char *zWhere, const char *zLabel){
     cnt++;
   }
   db_finalize(&q);
-  db_multi_exec("DELETE FROM description;");
+  if( zWhere!=0 ) db_multi_exec("DELETE FROM description;");
   return cnt;
 }
 
@@ -980,18 +1005,31 @@ void test_describe_artifacts_cmd(void){
 **
 **   n=N         Show N artifacts
 **   s=S         Start with artifact number S
+**   unpub       Show only unpublished artifacts
+**   hclr        Color code hash types (SHA1 vs SHA3)
 */
 void bloblist_page(void){
   Stmt q;
   int s = atoi(PD("s","0"));
   int n = atoi(PD("n","5000"));
   int mx = db_int(0, "SELECT max(rid) FROM blob");
+  int unpubOnly = PB("unpub");
+  int hashClr = PB("hclr");
   char *zRange;
+  char *zSha1Bg;
+  char *zSha3Bg;
 
   login_check_credentials();
   if( !g.perm.Read ){ login_needed(g.anon.Read); return; }
   style_header("List Of Artifacts");
-  if( mx>n && P("s")==0 ){
+  style_submenu_element("250 Largest", "bigbloblist");
+  if( g.perm.Admin ){
+    style_submenu_element("Artifact Log", "rcvfromlist");
+  }
+  if( g.perm.Write ){
+    style_submenu_element("Artifact Stats", "artifact_stats");
+  }
+  if( !unpubOnly && mx>n && P("s")==0 ){
     int i;
     @ <p>Select a range of artifacts to view:</p>
     @ <ul>
@@ -1003,22 +1041,39 @@ void bloblist_page(void){
     style_footer();
     return;
   }
-  if( mx>n ){
-    style_submenu_element("Index", "Index", "bloblist");
+  if( !unpubOnly && mx>n ){
+    style_submenu_element("Index", "bloblist");
   }
-  zRange = mprintf("BETWEEN %d AND %d", s, s+n-1);
+  if( unpubOnly ){
+    zRange = mprintf("IN private");
+  }else{
+    zRange = mprintf("BETWEEN %d AND %d", s, s+n-1);
+  }
   describe_artifacts(zRange);
+  fossil_free(zRange);
   db_prepare(&q,
     "SELECT rid, uuid, summary, isPrivate FROM description ORDER BY rid"
   );
+  if( skin_detail_boolean("white-foreground") ){
+    zSha1Bg = "#714417";
+    zSha3Bg = "#177117";
+  }else{
+    zSha1Bg = "#ebffb0";
+    zSha3Bg = "#b0ffb0";
+  }
   @ <table cellpadding="0" cellspacing="0">
   while( db_step(&q)==SQLITE_ROW ){
     int rid = db_column_int(&q,0);
     const char *zUuid = db_column_text(&q, 1);
     const char *zDesc = db_column_text(&q, 2);
-    int isPriv = db_column_int(&q,2);
-    @ <tr><td align="right">%d(rid)</td>
-    @ <td>&nbsp;%z(href("%R/info/%!S",zUuid))%s(zUuid)</a>&nbsp;</td>
+    int isPriv = db_column_int(&q,3);
+    if( hashClr ){
+      const char *zClr = db_column_bytes(&q,1)>40 ? zSha3Bg : zSha1Bg;
+      @ <tr style='background-color:%s(zClr);'><td align="right">%d(rid)</td>
+    }else{
+      @ <tr><td align="right">%d(rid)</td>
+    }
+    @ <td>&nbsp;%z(href("%R/info/%!S",zUuid))%S(zUuid)</a>&nbsp;</td>
     @ <td align="left">%h(zDesc)</td>
     if( isPriv ){
       @ <td>(unpublished)</td>
@@ -1027,6 +1082,70 @@ void bloblist_page(void){
   }
   @ </table>
   db_finalize(&q);
+  style_footer();
+}
+
+/*
+** WEBPAGE: bigbloblist
+**
+** Return a page showing the largest artifacts in the repository in order
+** of decreasing size.
+**
+**   n=N         Show the top N artifacts
+*/
+void bigbloblist_page(void){
+  Stmt q;
+  int n = atoi(PD("n","250"));
+
+  login_check_credentials();
+  if( !g.perm.Read ){ login_needed(g.anon.Read); return; }
+  if( g.perm.Admin ){
+    style_submenu_element("Artifact Log", "rcvfromlist");
+  }
+  if( g.perm.Write ){
+    style_submenu_element("Artifact Stats", "artifact_stats");
+  }
+  style_submenu_element("All Artifacts", "bloblist");
+  style_header("%d Largest Artifacts", n);
+  db_multi_exec(
+    "CREATE TEMP TABLE toshow(rid INTEGER PRIMARY KEY);"
+    "INSERT INTO toshow(rid)"
+    "  SELECT rid FROM blob"
+    "   ORDER BY length(content) DESC"
+    "   LIMIT %d;", n
+  );
+  describe_artifacts("IN toshow");
+  db_prepare(&q,
+    "SELECT description.rid, description.uuid, description.summary,"
+    "       length(blob.content), coalesce(delta.srcid,''),"
+    "       datetime(description.ctime)"
+    "  FROM description, blob LEFT JOIN delta ON delta.rid=blob.rid"
+    " WHERE description.rid=blob.rid"
+    " ORDER BY length(content) DESC"
+  );
+  @ <table cellpadding="2" cellspacing="0" border="1" \
+  @  class='sortable' data-column-types='NnnttT' data-init-sort='0'>
+  @ <thead><tr><th align="right">Size<th align="right">RID
+  @ <th align="right">Delta From<th>Hash<th>Description<th>Date</tr></thead>
+  @ <tbody>
+  while( db_step(&q)==SQLITE_ROW ){
+    int rid = db_column_int(&q,0);
+    const char *zUuid = db_column_text(&q, 1);
+    const char *zDesc = db_column_text(&q, 2);
+    int sz = db_column_int(&q,3);
+    const char *zSrcId = db_column_text(&q,4);
+    const char *zDate = db_column_text(&q,5);
+    @ <tr><td align="right">%d(sz)</td>
+    @ <td align="right">%d(rid)</td>
+    @ <td align="right">%s(zSrcId)</td>
+    @ <td>&nbsp;%z(href("%R/info/%!S",zUuid))%S(zUuid)</a>&nbsp;</td>
+    @ <td align="left">%h(zDesc)</td>
+    @ <td align="left">%z(href("%R/timeline?c=%T",zDate))%s(zDate)</a></td>
+    @ </tr>
+  }
+  @ </tbody></table>
+  db_finalize(&q);
+  style_table_sorter();
   style_footer();
 }
 
@@ -1070,19 +1189,19 @@ void test_phatoms_cmd(void){
 #define MAX_COLLIDE 25
 
 /*
-** Generate a report on the number of collisions in SHA1 hashes
+** Generate a report on the number of collisions in artifact hashes
 ** generated by the SQL given in the argument.
 */
 static void collision_report(const char *zSql){
   int i, j, kk;
   int nHash = 0;
   Stmt q;
-  char zPrev[UUID_SIZE+1];
+  char zPrev[HNAME_MAX+1];
   struct {
     int cnt;
     char *azHit[MAX_COLLIDE];
-    char z[UUID_SIZE+1];
-  } aCollide[UUID_SIZE+1];
+    char z[HNAME_MAX+1];
+  } aCollide[HNAME_MAX+1];
   memset(aCollide, 0, sizeof(aCollide));
   memset(zPrev, 0, sizeof(zPrev));
   db_prepare(&q,"%s",zSql/*safe-for-%s*/);
@@ -1092,7 +1211,7 @@ static void collision_report(const char *zSql){
     int i;
     nHash++;
     for(i=0; zPrev[i] && zPrev[i]==zUuid[i]; i++){}
-    if( i>0 && i<=UUID_SIZE ){
+    if( i>0 && i<=HNAME_MAX ){
       if( i>=4 && aCollide[i].cnt<MAX_COLLIDE ){
         aCollide[i].azHit[aCollide[i].cnt] = mprintf("%.*s", i, zPrev);
       }
@@ -1105,14 +1224,14 @@ static void collision_report(const char *zSql){
   @ <table border=1><thead>
   @ <tr><th>Length<th>Instances<th>First Instance</tr>
   @ </thead><tbody>
-  for(i=1; i<=UUID_SIZE; i++){
+  for(i=1; i<=HNAME_MAX; i++){
     if( aCollide[i].cnt==0 ) continue;
     @ <tr><td>%d(i)<td>%d(aCollide[i].cnt)<td>%h(aCollide[i].z)</tr>
   }
   @ </tbody></table>
   @ <p>Total number of hashes: %d(nHash)</p>
   kk = 0;
-  for(i=UUID_SIZE; i>=4; i--){
+  for(i=HNAME_MAX; i>=4; i--){
     if( aCollide[i].cnt==0 ) continue;
     if( aCollide[i].cnt>200 ) break;
     kk += aCollide[i].cnt;
@@ -1127,7 +1246,7 @@ static void collision_report(const char *zSql){
       @ %z(href("%R/whatis/%s",zId))%h(zId)</a>
     }
   }
-  for(i=4; i<ArraySize(aCollide); i++){
+  for(i=4; i<count(aCollide); i++){
     for(j=0; j<aCollide[i].cnt && j<MAX_COLLIDE; j++){
       fossil_free(aCollide[i].azHit[j]);
     }
@@ -1142,9 +1261,9 @@ static void collision_report(const char *zSql){
 void hash_collisions_webpage(void){
   login_check_credentials();
   if( !g.perm.Read ){ login_needed(g.anon.Read); return; }
-  style_header("SHA1 Prefix Collisions");
-  style_submenu_element("Activity Reports", 0, "reports");
-  style_submenu_element("Stats", 0, "stat");
+  style_header("Hash Prefix Collisions");
+  style_submenu_element("Activity Reports", "reports");
+  style_submenu_element("Stats", "stat");
   @ <h1>Hash Prefix Collisions on Check-ins</h1>
   collision_report("SELECT (SELECT uuid FROM blob WHERE rid=objid)"
                    "  FROM event WHERE event.type='ci'"
